requirements.txt
streamlit
openai
ollama
textattack
deep-translator
python-dotenv